user_identification_module/requirements.txt
boto3==1.34.144
opencv-python==4.8.1.78
Pillow==10.0.0
python-dotenv==1.0.0
Flask==2.3.3
Flask-SocketIO==5.3.6
numpy<2.0
pymongo==4.6.0
face-recognition==1.3.0
scikit-learn==1.3.2